test: add test-plugin-utils
