+++ packages/starlight/__tests__/plugins/integration.test.ts
@@ -1,7 +1,7 @@
 import type { AstroIntegration } from 'astro';
 import { expect, test } from 'vitest';
 import { runPlugins } from '../../utils/plugins';
-import { createTestPluginContext } from '../test-config';
+import { createTestPluginContext } from '../test-plugin-utils';
 
 test('returns all integrations added by plugins without deduping them', async () => {
 	const integration1: AstroIntegration = {
